Outsource PickUserName to src/components
